had to remove rockyou.txt from this
rockyou.txt was a txt file with like 14 mil most common passwords, and well that took up like 130 MB of space, and github didn't like that so... it had to go. you can still get a copy of it for some of my code. its not hard to find you literally just need to google the name of the file, its literally a famous dictionary attack wordlist.